Use More OOP
