+++ src/main/java/alpaca/Alpaca.java
@@ -1,0 +1,18 @@
+package alpaca;
+
+import Ui;
+
+public class Alpaca {
+    private Ui ui;
+
+    public Alpaca() {
+        ui = new Ui();
+    }
+
+    private void run() {
+        ui.run();
+    }
+    public static void main(String[] args) {
+        new Alpaca().run();
+    }
+}
