+++ src/main/java/alpaca/exceptions/InvalidInput.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package alpaca.exceptions;
 
 public class InvalidInput extends Exception {
     public InvalidInput(String errorMessage) {
