+++ src/main/java/alpaca/exceptions/ValueNotFound.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package alpaca.exceptions;
 
 public class ValueNotFound extends Exception {
     public ValueNotFound(String errorMessage) {
