+++ src/main/java/alpaca/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package Tasks;
+package alpaca.tasks;
 
 public class Deadline extends Task{
     public Deadline(String name) {
