+++ src/main/java/alpaca/tasks/Event.java
@@ -1,4 +1,4 @@
-package Tasks;
+package alpaca.tasks;
 
 public class Event extends Task{
     public Event(String name) {
