+++ src/main/java/alpaca/tasks/Task.java
@@ -1,4 +1,4 @@
-package Tasks;
+package alpaca.tasks;
 
 import java.util.ArrayList;
 import java.util.regex.Matcher;
