+++ src/main/java/alpaca/tasks/ToDo.java
@@ -1,4 +1,4 @@
-package Tasks;
+package alpaca.tasks;
 
 public class ToDo extends Task {
     public ToDo(String name) {
